Stripping/Rebuilding the build-system
Stripped out the original buildsystem, transitioning from the original C
implementation of Kconfig taken from linux to a python implemenation to
help simplify the build system. Overall just meant to get rid of most of
the small changes made indepentantly to the build system from multiple
different independant projects. Should also make it easier to integrate
LLVM compiler passes with the buildsystem.

+++ src/arch/arm64/iomap.c
@@ -15,7 +15,7 @@ int arch_handle_io_map(struct nk_io_mapping *mapping)
     return 0;
   } else {
     int flags = atomic_or(mapping->flags, NK_IO_MAPPING_FLAG_ENABLED);
-    return pt_init_table_device(ttbr0_table, mapping->vaddr, mapping->paddr, mapping->size);
+    return pt_init_table_device(ttbr0_table, (uint64_t)mapping->vaddr, (uint64_t)mapping->paddr, mapping->size);
   }
 }
 int arch_handle_io_unmap(struct nk_io_mapping *mapping)
